// types/streamType.tsx
export interface streamsType {
    course_id: number | null;
    stream_id: number;
    info: string | null;
    id_curricula: number;
    id_subject: number;
    subject_type: string;
    id_subject_type: number;
    id_edu_year: number;
    id_period: number;
    id_speciality: number;
    id_faculty: number;
    stream_title:string; 
}